Added Pet Text Fields, Sort, Add, Remove, Adopt Buttons, Adopt Functionality, Add, Remove Functionality
Added text fields to view information of selected pets in the list, as
well as to enter and create new pets to be added to the list.
Added functionality to remove Pets from the list of pets in the shelter
by selecting a pet from the list and removing it.
Added ability to sort list by name, age, and species.
Added ability to adopt pets, a label will be displayed to represent the
status of the pet (adopted or not), if an attempt is made to adopt a pet
that has already been adopted, a message will display indicating that
the pet has already been adopted.